Adds a dedicated strategy for the Sandbox API
Why:

* When developing apps on top of this gem, weird errors would occur
depending on how `ENV` was loaded.
The original approach depended on the fact that `ENV['UPHOLD_URL']` was
defined before the strategy was loaded. This works when using `foreman`
(which loads `.env` before starting Ruby) but not when using
`dotenv-rails`, which would only load the classes have already been
loaded. This would result in OmniAuth ignoring the ENV settings.

This change addresses the need by:

* Moving the sandbox URLs to a dedicated strategy, and turning constants
into class variables.